Move different resume sections into their own submenus
